idlib: drop redundant INFINITY guards; use stdint for temp-file counter
Tidy-up now that the real INFINITY fix lives at the use sites (Bounds.h,
Box.h). The two earlier defensive undefs added while diagnosing the macOS
failure are redundant and removed:
 - the end-of-Math.h undef (couldn't help: Math.h's include guard stops it
   re-running after <float.h> redefines INFINITY later in the TU)
 - the platform.h post-<math.h> undef (wrong header: the macro that bit us
   came from <float.h>, and the collision is at the use site regardless)
The original top-of-Math.h undef is kept (pre-existing, harmless).

Also switch the MakeTemporaryFile counter from int to uint32_t (resolves
via platform.h's <inttypes.h>) per the request to prefer stdint integer
types. The VFS flag helper keeps 'unsigned' deliberately - that matches
filestream_open()'s parameter type, so stdint there would only introduce
a mismatch.

Verified: Linux build unchanged; the exact macOS "expected unqualified-id"
error stays fixed by the use-site undef with these guards removed.

+++ neo/framework/FileSystem.cpp
@@ -3817,10 +3817,10 @@ idFileSystemLocal::MakeTemporaryFile
 idFile * idFileSystemLocal::MakeTemporaryFile( void ) {
 	// libretro-common has no tmpfile() analog; create a uniquely-named
 	// scratch file under the save path and open it read/write.
-	static int tmpCounter = 0;
+	static uint32_t tmpCounter = 0;
 	idStr tmpPath = fs_savepath.GetString();
 	tmpPath += PATHSEPERATOR_STR;
-	tmpPath += va( "d3tmp_%u_%d.tmp", (unsigned)Core_Milliseconds(), tmpCounter++ );
+	tmpPath += va( "d3tmp_%u_%u.tmp", (unsigned)(uint32_t)Core_Milliseconds(), (unsigned)tmpCounter++ );
 	RFILE *f = filestream_open( tmpPath.c_str(),
 		RETRO_VFS_FILE_ACCESS_READ_WRITE,
 		RETRO_VFS_FILE_ACCESS_HINT_NONE );
